requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
einops==0.8.1
eval_type_backport==0.3.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gymnasium==1.1.1
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.7
lightning==2.6.0
lightning-utilities==0.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
openai==2.8.1
orjson==3.11.4
ortools==9.14.6206
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.31.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
robust-downloader==0.0.2
scipy==1.13.1
sentry-sdk==2.46.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tensordict==0.10.0

torchmetrics==1.8.2

tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
yarl==1.22.0
zipp==3.23.0
